add exp backoff before delete project instead of sleep
