+++ tests/core/GHC-ORG-PRI-UNA.js
@@ -269,13 +269,7 @@ describe(testSuite + testSuiteDesc,
             assert.strictEqual(err, 401, util.format('Cannot cancel build  ' +
               'id: %d for project id: %s, err: %s, %s', runId, projectId, err,
               response));
-            logger.info(util.format('sleeping %s ms after cancel',
-              global.DELETE_PROJ_DELAY));
-            setTimeout(
-              function () {
-                return done();
-              }, global.DELETE_PROJ_DELAY
-            );
+            return done();
           }
         );
       }
@@ -344,27 +338,7 @@ describe(testSuite + testSuiteDesc,
     after(
       function (done) {
         if (projectId)
-          global.suAdapter.deleteProjectById(projectId, {},
-            function (err, response) {
-              if (err) {
-                logger.warn(testSuite,
-                  util.format('Cleanup-failed to delete the project with id:' +
-                    '%s, err: %s, %s', projectId, err, util.inspect(response)
-                  )
-                );
-                return done();
-              }
-              global.removeResource(
-                {
-                  type: 'project',
-                  id: projectId
-                },
-                function () {
-                  return done();
-                }
-              );
-            }
-          );
+          global.deleteProjectWithBackoff(projectId, done);
         else
           return done();
       }
